+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ### Added
 
+- ✨(plugins) add endpoint to list siret of active organizations #771
 - ✨(core) create AccountServiceAuthentication backend #771
 - ✨(core) create AccountService model #771
 - 🔒️(passwords) add validators for production #850